mcp-server/requirements.txt
mcp
uvicorn
opensearch-py
python-dotenv
langchain
langchain-aws
langchain-openai
langchain-community
boto3
botocore
requests-aws4auth
tiktoken
httpx